BasicESC support. Change ESC interface to double (rangle -1 to 1)

+++ src/main/java/net/catchpole/B9/devices/esc/ESC.java
@@ -3,5 +3,5 @@ package net.catchpole.B9.devices.esc;
 import java.io.IOException;
 
 public interface ESC {
-    void update(int throttle) throws IOException;
+    void update(double throttle) throws IOException;
 }
